% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateConductance.R
\name{calculateConductance}
\alias{calculateConductance}
\title{Calculate conductance}
\usage{
calculateConductance(
  shotData,
  flowColumn = "espresso_flow",
  pressureColumn = "espresso_pressure",
  conductanceColumnName = "espresso_conductance"
)
}
\arguments{
\item{shotData}{numeric matrix from .shot file, containing pressure and flow data}

\item{flowColumn}{character string of column pointing to flow data; default='espresso_flow'}

\item{pressureColumn}{character string of column pointing to pressure data; default='espresso_pressure'}

\item{conductanceColumnName}{character string of resulting column name that will be appended with conductance data; default='espresso_conductance'}
}
\value{
original shotData matrix with appended conductance values with name conductanceColumnName
}
\description{
Calculate conductance given pressure and flow data, generated from .shot files.
}
\details{
Conductance = (Flow^2)/Pressure
}
\examples{
tmp1 <- importShotFile(paste0(path,file))
dim(head(tmp1))
[1]  6 12
tmp1 <- calculateConductance(tmp1)
dim(head(tmp1))
[1]  6 13

}
\references{
Collin Arneson
}
\author{
Erol Biceroglu
}
