% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateResistance.R
\name{calculateResistance}
\alias{calculateResistance}
\title{Calculate Resistance}
\usage{
calculateResistance(
  shotData,
  flowColumn = "espresso_flow",
  pressureColumn = "espresso_pressure",
  resistanceColumnName = "espresso_resistance"
)
}
\arguments{
\item{shotData}{numeric matrix from .shot file, containing pressure and flow data}

\item{flowColumn}{character string of column pointing to flow data; default='espresso_flow'}

\item{pressureColumn}{character string of column pointing to pressure data; default='espresso_pressure'}

\item{resistanceColumnName}{character string of resulting column name that will be appended with resistance data; default='espresso_resistance'}
}
\value{
original shotData matrix with appended resistance values with name resistanceColumnName
}
\description{
Calculate resistance given pressure and flow data, generated from .shot files.
}
\details{
Resistance = Pressure/(Flow^2)
}
\examples{
tmp1 <- importShotFile(paste0(path,file))
dim(head(tmp1))
[1]  6 12
tmp1 <- calculateResistance(tmp1)
dim(head(tmp1))
[1]  6 13

}
\references{
Collin Arneson
}
\author{
Erol Biceroglu
}
